// src/generated/model/collectPaymentRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CollectPaymentRequestPaymentMethod } from './collectPaymentRequestPaymentMethod.ts';

export interface CollectPaymentRequest {
  paymentDate: string;
  paymentMethod: CollectPaymentRequestPaymentMethod;
  amount: number;
  transactionReference?: string;
  bankName?: string;
  accountName?: string;
  accountNumber?: string;
  referenceNumber?: string;
  remarks?: string;
}