// src/generated/model/filterCriteria.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FilterCriteriaOperator } from './filterCriteriaOperator.ts';

export interface FilterCriteria {
  field?: string;
  operator?: FilterCriteriaOperator;
  value?: unknown;
}